+++ src/it/test-projects/creedengo-python-plugin-test-project/src/usingSlotsOnDataClassesNonCompliant.py
@@ -1,17 +1,23 @@
-@dataclass
-class MyClass1:
+@dataclass # Noncompliant
+class MyClass:
     a: int
     b: int
     c: int
 
-@dataclass()
+@dataclass() # Noncompliant
 class MyClass2:
     a: int
     b: int
     c: int
 
-@dataclass(frozen=True)
+@dataclass(frozen=True) # Noncompliant
 class MyClass3:
+    a: int
+    b: int
+    c: int
+
+@dataclass(slots=False) # Noncompliant
+class MyClass4:
     a: int
     b: int
     c: int